# PersonDetails/PersonDetails/api.yaml
openapi: 3.0.3
info:
  title: User CRUD API
  version: 1.0.0
  description: API for managing users

servers:
  - url: http://localhost:8081/api

paths:

  /users:

    get:
      summary: Get all users (with pagination)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0

        - name: size
          in: query
          schema:
            type: integer
            default: 10

      responses:
        '200':
          description: Users fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /users/{id}:

    get:
      summary: Get user by ID
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      summary: Delete user
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:

  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

  schemas:

    UserRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: Akhib Khan

        email:
          type: string
          format: email

        phone:
          type: string
          example: "9876543210"

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string

    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserResponse'

    PaginatedResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: ERROR
        message:
          type: string
        errorCode:
          type: string