# PersonDetails/PersonDetails/src/main/java/com/example/PersonDetails/dto/UserResponse.java
package com.example.PersonDetails.dto;

public class UserResponse {

    private String id;
    private String name;
    private String email;
    private String phone;

    public String getId() { return id; }
    public void setId(String id) { this.id = id; }

    public String getName() { return name; }
    public void setName(String name) { this.name = name; }

    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }

    public String getPhone() { return phone; }
    public void setPhone(String phone) { this.phone = phone; }
}